split dictionary into multiple files
