+++ src/libnetdata/dictionary/dictionary.h
@@ -328,14 +328,4 @@ extern struct dictionary_stats dictionary_stats_category_other;
 
 int dictionary_unittest(size_t entries);
 
-// ----------------------------------------------------------------------------
-// THREAD CACHE
-
-void *thread_cache_entry_get_or_set(void *key,
-                                    ssize_t key_length,
-                                    void *value,
-                                    void *(*transform_the_value_before_insert)(void *key, size_t key_length, void *value));
-
-void thread_cache_destroy(void);
-
 #endif /* NETDATA_DICTIONARY_H */
